requirements.txt
Flask==2.0.1
openai==0.27.0
python-dotenv==0.20.0
flask-cors==3.0.10
requests==2.27.1